restructure file format
